🚚 Move NewVersionModal to /documents/modals
