+++ src/documents/modals/NewVersionModal/UploadStep.js
