+++ src/documents/modals/NewVersionModal/__tests__/NewVersionFlow.test.js
@@ -4,8 +4,8 @@ import {MockedProvider} from 'react-apollo/test-utils';
 import {MemoryRouter} from 'react-router-dom';
 import {render, fireEvent, cleanup, act} from 'react-testing-library';
 import {NewVersionFlow} from '../NewVersionFlow';
-import {mocks} from '../../../../__mocks__/kf-api-study-creator/mocks';
-import fileByKfId from '../../../../__mocks__/kf-api-study-creator/responses/fileByKfId.json';
+import {mocks} from '../../../../../__mocks__/kf-api-study-creator/mocks';
+import fileByKfId from '../../../../../__mocks__/kf-api-study-creator/responses/fileByKfId.json';
 
 afterEach(cleanup);
